+++ cosim4j-native/src/jni/execution.cpp
@@ -11,7 +11,7 @@ extern "C" {
 #endif
 
 
-JNIEXPORT jlong JNICALL Java_org_osp_cosim_jni_CosimLibrary_createExecution(JNIEnv* env, jobject, jdouble startTime, jdouble stepSize)
+JNIEXPORT jlong JNICALL Java_com_opensimulationplatform_cosim_jni_CosimLibrary_createExecution(JNIEnv* env, jobject, jdouble startTime, jdouble stepSize)
 {
     cosim_execution* execution = cosim_execution_create(to_cosim_time_point(startTime), to_cosim_duration(stepSize));
     if (execution == nullptr) {
@@ -22,7 +22,7 @@ JNIEXPORT jlong JNICALL Java_org_osp_cosim_jni_CosimLibrary_createExecution(JNIE
     return reinterpret_cast<jlong>(execution);
 }
 
-JNIEXPORT jlong JNICALL Java_org_osp_cosim_jni_CosimLibrary_createSspExecution(JNIEnv* env, jobject, jstring sspPath, jboolean startTimeDefined, jdouble startTime)
+JNIEXPORT jlong JNICALL Java_com_opensimulationplatform_cosim_jni_CosimLibrary_createSspExecution(JNIEnv* env, jobject, jstring sspPath, jboolean startTimeDefined, jdouble startTime)
 {
     auto sspDir_ = env->GetStringUTFChars(sspPath, nullptr);
     cosim_execution* execution = cosim_ssp_execution_create(sspDir_, startTimeDefined, to_cosim_time_point(startTime));
@@ -35,7 +35,7 @@ JNIEXPORT jlong JNICALL Java_org_osp_cosim_jni_CosimLibrary_createSspExecution(J
     return reinterpret_cast<jlong>(execution);
 }
 
-JNIEXPORT jlong JNICALL Java_org_osp_cosim_jni_CosimLibrary_createOspConfigExecution(JNIEnv* env, jobject, jstring configPath, jboolean startTimeDefined, jdouble startTime)
+JNIEXPORT jlong JNICALL Java_com_opensimulationplatform_cosim_jni_CosimLibrary_createOspConfigExecution(JNIEnv* env, jobject, jstring configPath, jboolean startTimeDefined, jdouble startTime)
 {
     auto configDir_ = env->GetStringUTFChars(configPath, nullptr);
     cosim_execution* execution = cosim_osp_config_execution_create(configDir_, startTimeDefined, to_cosim_time_point(startTime));
@@ -48,7 +48,7 @@ JNIEXPORT jlong JNICALL Java_org_osp_cosim_jni_CosimLibrary_createOspConfigExecu
     return reinterpret_cast<jlong>(execution);
 }
 
-JNIEXPORT jlong JNICALL Java_org_osp_cosim_jni_CosimLibrary_createFixedStepSspExecution(JNIEnv* env, jobject, jstring sspDir, jdouble stepSize, jboolean startTimeDefined, jdouble startTime)
+JNIEXPORT jlong JNICALL Java_com_opensimulationplatform_cosim_jni_CosimLibrary_createFixedStepSspExecution(JNIEnv* env, jobject, jstring sspDir, jdouble stepSize, jboolean startTimeDefined, jdouble startTime)
 {
     auto sspDir_ = env->GetStringUTFChars(sspDir, nullptr);
     cosim_execution* execution = cosim_ssp_fixed_step_execution_create(
@@ -65,7 +65,7 @@ JNIEXPORT jlong JNICALL Java_org_osp_cosim_jni_CosimLibrary_createFixedStepSspEx
     return reinterpret_cast<jlong>(execution);
 }
 
-JNIEXPORT jboolean JNICALL Java_org_osp_cosim_jni_CosimLibrary_destroyExecution(JNIEnv* env, jobject, jlong executionPtr)
+JNIEXPORT jboolean JNICALL Java_com_opensimulationplatform_cosim_jni_CosimLibrary_destroyExecution(JNIEnv* env, jobject, jlong executionPtr)
 {
     if (executionPtr == 0) {
         std::cerr << "[JNI-wrapper] Error: executionPtr is NULL" << std::endl;
@@ -75,27 +75,27 @@ JNIEXPORT jboolean JNICALL Java_org_osp_cosim_jni_CosimLibrary_destroyExecution(
 }
 
 
-JNIEXPORT jboolean JNICALL Java_org_osp_cosim_jni_CosimLibrary_start(JNIEnv* env, jobject, jlong executionPtr)
+JNIEXPORT jboolean JNICALL Java_com_opensimulationplatform_cosim_jni_CosimLibrary_start(JNIEnv* env, jobject, jlong executionPtr)
 {
     return cosim_execution_start(reinterpret_cast<cosim_execution*>(executionPtr)) == 0;
 }
 
-JNIEXPORT jboolean JNICALL Java_org_osp_cosim_jni_CosimLibrary_step(JNIEnv* env, jobject, jlong executionPtr, jlong numSteps)
+JNIEXPORT jboolean JNICALL Java_com_opensimulationplatform_cosim_jni_CosimLibrary_step(JNIEnv* env, jobject, jlong executionPtr, jlong numSteps)
 {
     return cosim_execution_step(reinterpret_cast<cosim_execution*>(executionPtr), numSteps) == 0;
 }
 
-JNIEXPORT jboolean JNICALL Java_org_osp_cosim_jni_CosimLibrary_simulateUntil(JNIEnv* env, jobject, jlong executionPtr, jdouble targetTime)
+JNIEXPORT jboolean JNICALL Java_com_opensimulationplatform_cosim_jni_CosimLibrary_simulateUntil(JNIEnv* env, jobject, jlong executionPtr, jdouble targetTime)
 {
     return cosim_execution_simulate_until(reinterpret_cast<cosim_execution*>(executionPtr), to_cosim_time_point(targetTime)) == 0;
 }
 
-JNIEXPORT jboolean JNICALL Java_org_osp_cosim_jni_CosimLibrary_stop(JNIEnv* env, jobject, jlong executionPtr)
+JNIEXPORT jboolean JNICALL Java_com_opensimulationplatform_cosim_jni_CosimLibrary_stop(JNIEnv* env, jobject, jlong executionPtr)
 {
     return cosim_execution_stop(reinterpret_cast<cosim_execution*>(executionPtr)) == 0;
 }
 
-JNIEXPORT jboolean JNICALL Java_org_osp_cosim_jni_CosimLibrary_enableRealTimeSimulation(JNIEnv* env, jobject, jlong executionPtr)
+JNIEXPORT jboolean JNICALL Java_com_opensimulationplatform_cosim_jni_CosimLibrary_enableRealTimeSimulation(JNIEnv* env, jobject, jlong executionPtr)
 {
     if (executionPtr == 0) {
         std::cerr << "[JNI-wrapper] Error: executionPtr is NULL" << std::endl;
@@ -104,7 +104,7 @@ JNIEXPORT jboolean JNICALL Java_org_osp_cosim_jni_CosimLibrary_enableRealTimeSim
     return cosim_execution_enable_real_time_simulation(reinterpret_cast<cosim_execution*>(executionPtr)) == 0;
 }
 
-JNIEXPORT jboolean JNICALL Java_org_osp_cosim_jni_CosimLibrary_disableRealTimeSimulation(JNIEnv* env, jobject, jlong executionPtr)
+JNIEXPORT jboolean JNICALL Java_com_opensimulationplatform_cosim_jni_CosimLibrary_disableRealTimeSimulation(JNIEnv* env, jobject, jlong executionPtr)
 {
     if (executionPtr == 0) {
         std::cerr << "[JNI-wrapper] Error: executionPtr is NULL" << std::endl;
@@ -113,7 +113,7 @@ JNIEXPORT jboolean JNICALL Java_org_osp_cosim_jni_CosimLibrary_disableRealTimeSi
     return cosim_execution_disable_real_time_simulation(reinterpret_cast<cosim_execution*>(executionPtr)) == 0;
 }
 
-JNIEXPORT jboolean JNICALL Java_org_osp_cosim_jni_CosimLibrary_setRealTimeFactorTarget(JNIEnv* env, jobject, jlong executionPtr, jdouble realTimeFactor)
+JNIEXPORT jboolean JNICALL Java_com_opensimulationplatform_cosim_jni_CosimLibrary_setRealTimeFactorTarget(JNIEnv* env, jobject, jlong executionPtr, jdouble realTimeFactor)
 {
     if (executionPtr == 0) {
         std::cerr << "[JNI-wrapper] Error: executionPtr is NULL" << std::endl;
@@ -122,7 +122,7 @@ JNIEXPORT jboolean JNICALL Java_org_osp_cosim_jni_CosimLibrary_setRealTimeFactor
     return cosim_execution_set_real_time_factor_target(reinterpret_cast<cosim_execution*>(executionPtr), realTimeFactor) == 0;
 }
 
-JNIEXPORT jobject JNICALL Java_org_osp_cosim_jni_CosimLibrary_getStatus(JNIEnv* env, jobject, jlong executionPtr)
+JNIEXPORT jobject JNICALL Java_com_opensimulationplatform_cosim_jni_CosimLibrary_getStatus(JNIEnv* env, jobject, jlong executionPtr)
 {
     if (executionPtr == 0) {
         std::cerr << "[JNI-wrapper] Error: executionPtr is NULL" << std::endl;
